pyproject.toml
[project]
name = "underwriterdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "google-adk>=1.28.0",
    "google-cloud-bigquery>=3.40.1",
    "google-genai>=1.68.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
